+++ src/windows/editor_object_selection.c
@@ -846,7 +846,7 @@ static void window_editor_object_selection_mouseup(rct_window *w, int widgetInde
 		break;
 	case WIDX_FILTER_STRING_BUTTON:
 		//window_text_input_open(w, widgetIndex, STR_OBJECT_SEARCH, STR_OBJECT_SEARCH_DESC, STR_STRING, (uint32)_filter_string, 40);
-		window_start_textbox(w, widgetIndex, STR_STRING, (uint32)_filter_string, 40);
+		window_start_textbox(w, widgetIndex, STR_STRING, _filter_string, 40);
 		break;
 	case WIDX_FILTER_CLEAR_BUTTON:
 		memset(_filter_string, 0, sizeof(_filter_string));
@@ -1694,7 +1694,7 @@ static void window_editor_object_selection_manage_tracks()
 	gS6Info->editor_step = EDITOR_STEP_TRACK_DESIGNS_MANAGER;
 
 	int entry_index = 0;
-	for (; ((int)object_entry_groups[0].chunks[entry_index]) == -1; ++entry_index);
+	for (; ((intptr_t)object_entry_groups[0].chunks[entry_index]) == -1; ++entry_index);
 
 	rct_ride_entry* ride_entry = get_ride_entry(entry_index);
 	uint8* ride_type_array = &ride_entry->ride_type[0];
