+++ src/windows/server_list.c
@@ -210,7 +210,7 @@ static void window_server_list_mouseup(rct_window *w, int widgetIndex)
 		window_close(w);
 		break;
 	case WIDX_PLAYER_NAME_INPUT:
-		window_start_textbox(w, widgetIndex, STR_STRING, (uint32)_playerName, 63);
+		window_start_textbox(w, widgetIndex, STR_STRING, _playerName, 63);
 		break;
 	case WIDX_LIST:{
 		int serverIndex = w->selected_list_item;
